src/requirements.txt
aiohttp==3.7.4.post0
asgiref==3.5.2
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.1.3
colorama==0.4.4
discord.py==1.7.3
distlib==0.3.5
Django==4.1
django-ckeditor==6.5.0
django-js-asset==2.0.0
filelock==3.8.0
Flask==2.1.3
gTTS==2.2.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
multidict==5.2.0
nose==1.3.7
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
playsound==1.3.0
pycparser==2.21
pyenchant==3.2.2
PyNaCl==1.4.0
pyparsing==3.0.9
pytesseract==0.3.9
python-docx==0.8.11
python-dotenv==0.19.2
Random-Word==1.0.7
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.0
urllib3==1.26.8
virtualenv==20.16.3
Werkzeug==2.2.0
yarl==1.7.2
youtube-dl==2021.6.6
